Moving LSA to detail
