+++ app/demo-loop/diagnostic/EnergyDiagnostic.hh
@@ -19,8 +19,8 @@
 #include "corecel/data/CollectionBuilder.hh"
 #include "corecel/data/CollectionMirror.hh"
 #include "corecel/math/Atomics.hh"
-#include "orange/LevelStateAccessor.hh"
 #include "orange/Types.hh"
+#include "orange/detail/LevelStateAccessor.hh"
 #include "celeritas/Types.hh"
 #include "celeritas/global/CoreTrackData.hh"
 #include "celeritas/grid/NonuniformGrid.hh"
